requirements.txt
requests
kafka-python
apache-airflow==2.7.3
